api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django==5.2.8",
    "django-cors-headers>=4.9.0",
    "django-filter==25.2",
    "djangorestframework==3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "markdown==3.10",
]
